+++ recommendation-command-service/src/main/java/com/stackroute/squad/repository/SubDomainRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 /*@Repository annotation is used to indicate that the class provides the mechanism for storage, retrieval,
  search, update and delete operation on objects.*/
 @Repository
-public interface SubDomainRepository extends Neo4jRepository<SubDomain, Integer> {
+public interface SubDomainRepository extends Neo4jRepository<SubDomain,Long> {
   public SubDomain findById(int id);
   @Query("MATCH (sd:SubDomain)<-[h:has]-(d:domain) RETURN sd,h,d")
   public Domain domain();
